feat: Tags as part of <BookPage/> refactor: Tag Client-Side ustilities moves to /Taxonomy folder
