+++ src/app/(main)/book/create-or-sell-first/config.tsx
@@ -14,6 +14,7 @@ export const content = (
     factOne="Если можно продать без продукта, следует продавать."
     factTwo="Если без продукта вам не верят, надо создать его демонстрационные примеры."
     factThree="Если продукт не до конца определен, предварительные продажи могут помочь сориентироваться в потребностях рынка."
+    tags={tags}
   >
     <Typo>
       С одной стороны продавать «воздух» занятие не из простых. C другой стороны имея вагон «первоклассного чистого
