+++ src/app/(main)/book/doing-from-the-end/config.tsx
@@ -11,6 +11,7 @@ export const content = (
     headerLink={link}
     nextPage="/what-is-next/"
     quote="Решайте бизнес задачи в обратном порядке. Сначала сделайте все для приема денег и удовлетворения клиента, а уже потом можете заняться дистрибуцией и производством."
+    tags={tags}
   >
     <Typo>Если вы хотите продавать что-то на рынке то для этого надо (в обратном порядке):</Typo>
     <Typo variant="list">
