+++ src/app/(main)/book/layout.tsx
@@ -3,6 +3,7 @@ import { FunctionComponent, PropsWithChildren } from 'react';
 
 /**
  * Randers page with tags and categories
+ * @deprecated Please remove
  */
 const SinglePageLayout: FunctionComponent<PropsWithChildren> = ({ children }) => {
   return (
