+++ src/app/(main)/book/new-product-vs-existing-solution/config.tsx
@@ -11,6 +11,7 @@ export const content = (
     headerLink={link}
     nextPage="/business-as-1-10-100/"
     quote="Продавать свое так же сложно как продавать чужое, производить «собственный продукт» имеет смысл только если вы получаете ощутимое преимущество на рынке от этого."
+    tags={tags}
   >
     <Typo>
       Если вы все просчитали, определились с экономической целесообразностью, спланировали стратегию донесения продукта
