+++ src/components/Taxonomy/TagGroup.tsx
@@ -1,8 +1,8 @@
 import { FunctionComponent, HTMLAttributes } from 'react';
-import Typo from '../Typo';
-import Tag from './Tag';
-import styles from './Taxonomy.module.css';
 import Stack from '../Stack';
+import Tag from './Tag';
+import Typo from '../Typo';
+import styles from './Taxonomy.module.css';
 
 interface Props extends HTMLAttributes<HTMLDivElement> {
   tags?: string[];
